fix: created start module
